+++ ConsoleApp1/ConsoleApp1.csproj
@@ -5,6 +5,7 @@
     <TargetFramework>net8.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
+	  <LangVersion>12</LangVersion>  
   </PropertyGroup>
 
 </Project>
